src/app/extensions/cam-account/pages/account-profile/camfil-account-language-form/camfil-account-language-form.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { instance, mock } from 'ts-mockito';

import { AccountFacade } from 'ish-core/facades/account.facade';
import { CamfilErrorComponent } from 'ish-shared/components/common/camfil-error/camfil-error.component';

import { CamfilAccountLanguageFormComponent } from './camfil-account-language-form.component';

describe('Camfil Account Language Form Component', () => {
  let component: CamfilAccountLanguageFormComponent;
  let fixture: ComponentFixture<CamfilAccountLanguageFormComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [CamfilAccountLanguageFormComponent, CamfilErrorComponent],
      providers: [{ provide: AccountFacade, useFactory: () => instance(mock(AccountFacade)) }],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(CamfilAccountLanguageFormComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
